+++ include/nstd/gl/render.h
@@ -1,0 +1,6 @@
+#ifndef NSTD_GL_RENDER_H_INCLUDED
+#define NSTD_GL_RENDER_H_INCLUDED
+#include "../nstd.h"
+#include "render/pass.h"
+#include "render/pipeline.h"
+#endif