+++ include/nstd/gl/shader.h
@@ -1,0 +1,5 @@
+#ifndef NSTD_GL_SHADER_H_INCLUDED
+#define NSTD_GL_SHADER_H_INCLUDED
+#include "../nstd.h"
+#include "shader/module.h"
+#endif